Feature and bugfix update
Allowed Time Paradox system to switch player back to starting character if currently controlling another character, added ability to wash time machines in your garage, fixed bug in garage menu causing softlock if clicking empty screen space while in garage menu, fixed erroneous HP updates when entering/exiting RC Mode, made it so DeLorean can't stall if garage menu is open, fixed incorrect vehicle restart method when in RC Mode, prevented time machine map blips from disappearing when vehicle is no longer being rendered but still exists, prevented garage menu from being accessible while story missions are active, prevented Wayback and Time Paradox from activating while story missions are active

+++ BackToTheFutureV/Utility/Wayback/WaybackMachine.cs
@@ -148,7 +148,7 @@ namespace BackToTheFutureV
 
         private WaybackRecord Record()
         {
-            if ((IsPlayer && !FusionUtils.PlayerPed.IsAlive) || (!IsPlayer && !Ped.ExistsAndAlive()) || FusionUtils.CurrentTime < StartTime)
+            if ((IsPlayer && !FusionUtils.PlayerPed.IsAlive) || (!IsPlayer && !Ped.ExistsAndAlive()) || FusionUtils.CurrentTime < StartTime || Game.IsMissionActive)
             {
                 Stop();
                 return null;
@@ -180,8 +180,17 @@ namespace BackToTheFutureV
 
         private void Play()
         {
-            if ((Ped.NotNullAndExists() && Ped.IsDead && ModSettings.TimeParadox) || (Ped.NotNullAndExists() && Ped.LastVehicle.NotNullAndExists() && Ped.LastVehicle.IsConsideredDestroyed && ModSettings.TimeParadox))
+            if ((Ped.NotNullAndExists() && Ped.IsDead && !Game.IsMissionActive && ModSettings.TimeParadox) || (Ped.NotNullAndExists() && Ped.LastVehicle.NotNullAndExists() && Ped.LastVehicle.IsConsideredDestroyed && !Game.IsMissionActive && ModSettings.TimeParadox))
             {
+                if (FusionUtils.PlayerPed.CurrentVehicle.NotNullAndExists() && FusionUtils.PlayerPed.CurrentVehicle.IsTimeMachine() && TimeMachineHandler.CurrentTimeMachine.Properties.IsRemoteControlled)
+                {
+                    RemoteTimeMachineHandler.StopRemoteControl();
+                }
+                if (FusionUtils.PlayerPed.Model != Main.ResetPed.Model)
+                {
+                    Function.Call(Hash.CHANGE_PLAYER_PED, Game.Player, Main.SwitchedPed, false, false);
+                    PlayerSwitch.Disable = true;
+                }
                 FusionUtils.PlayerPed.Kill();
                 WaybackSystem.Paradox = true;
                 WaybackSystem.paradoxDelay = Game.GameTime + 3600;
